requirements.txt
nose
coverage
nose-cov
pytest
requests
pyyaml
six
h5py
matplotlib
numpy
scipy
scikit-learn
pandas
Pillow
imbalanced-learn
medpy
SimpleITK
